Add the schedule_to_start and schedule_to_close timeouts.
Those two timeouts are calculated based on the assumption that only one
worker per activity. So if the user has set the create of the activity
to be 10 minutes and a generator spawns 10 activities, the schedule to start
will be 100 minutes.

Schedule to close timeout is an addition of schedule to start and start to
close timeout.

@rantonmattei